requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
boto3==1.20.21
    # via -r requirements.in
botocore==1.23.21
    # via
    #   boto3
    #   s3transfer
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
python-dateutil==2.8.2
    # via botocore
pyyaml==6.0
    # via -r requirements.in
s3transfer==0.5.0
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.7
    # via botocore
